Crate headless_lms_server

The server that handles the requests.

See also

headless_lms_utils

headless_lms_models

Modules

config

Functionality for configuring the server

controllers

Handlers for HTTP requests to /api/v0.

domain

Helpful structs and functions that are not related to database tables.

prelude

Re-exports commonly used types for convenient use across the crate. Intended to be glob-imported like use crate::prelude::*;.

programs

Executable programs that can be started. Contains for example the server program, background services, and utility programs.

Macros

asset_json

Compile-time JSON inclusion macro. Usage: asset_json!("../../assets/quizzes.json")

doc

doc_path

generated_docs

Includes the type’s JSON example and/or TypeScript definition generated by doc-file-generator as a string. Used with the helper macro from the doc-macro crate: #generated_doc

Functions

setup_file_store

Setups file store so that it can be passed to actix web as data. Using Arc here so that this can be accessed from all the different worker threads.

setup_tracing

Sets up tokio tracing. Also makes sure that log statements from libraries respect the log level settings that have been set with RUST_LOG, for example:

Type Aliases

OAuthClient