headless_lms_server::controllers

Module auth

Handlers for HTTP requests to /api/v0/auth.

Structs

CreateAccountDetails

EmailCode

Login

SendEmailCodeData

UserInfo

Generic information about the logged in user.

Functions

_add_routes

authorize_action_on_resource

POST /api/v0/auth/authorize checks whether user can perform specified action on specified resource.

authorize_multiple_actions_on_resources

POST /api/v0/auth/authorize-multiple checks whether user can perform specified action on specified resource. Returns booleans for the authorizations in the same order as the input.

delete_user_account

POST /api/v0/auth/delete-user-account If users single-use code is correct then delete users account

handle_production_login 🔒

handle_test_mode_login 🔒

handle_test_mode_signup 🔒

handle_uuid_login 🔒

is_user_global_admin

logged_in

GET /api/v0/auth/logged-in Returns the current user’s login status.

login

POST /api/v0/auth/login Logs in to TMC. Returns true if login was successful, false if credentials were incorrect.

logout

POST /api/v0/auth/logout Logs out.

send_delete_user_email_code

POST /api/v0/auth/send-email-code If users password is correct, sends a code to users email for account deletion

signup

POST /api/v0/auth/signup Creates new mooc.fi account and signs in.

update_user_information_to_tmc

user_info

GET /api/v0/auth/user-info Returns the current user’s info.