headless_lms_server::controllers::cms

Module courses

Controllers for requests starting with /api/v0/cms/courses.

Functions

_add_routes

Add a route for each controller in this module.

add_media ğŸ”’

POST /api/v0/cms/courses/:course_id/upload - Uploads a media (image, audio, file) for the course from Gutenberg page edit.

delete_partners_block ğŸ”’

DELETE /api/v0/main-frontend/courses/:course_id - Delete a partners block in a course.

get_all_pages ğŸ”’

GET /api/v0/cms/courses/:course_id/pages - Gets all pages for a course.

get_course_by_id ğŸ”’

GET /api/v0/cms/courses/:course_id - Get the course.

get_course_default_peer_or_self_review_configuration ğŸ”’

get_course_instances ğŸ”’

GET /api/v0/cms/courses/:course_id/course-instances - Returns all course instances for given course id.

get_course_modules ğŸ”’

GET /api/v0/cms/courses/:course_id/modules Returns modules in the course.

get_course_nondefault_chatbot_configurations ğŸ”’

GET /api/v0/cms/courses/:course_id/nondefault-chatbot-configurations - Get all nondefault, enabled-to-students chatbot configurations of this course.

get_partners_block ğŸ”’

GET /courses/:course_id/partners_blocks - Gets a partners block related to a course

get_research_form_with_course_id ğŸ”’

GET /api/v0/cms/courses/:course_id/research-consent-form - Fetches courses research form with course id.

post_partners_block ğŸ”’

POST /api/v0/main-frontend/courses/:course_id/partners_block - Create or updates a partners block for a course

put_course_default_peer_or_self_review_configuration ğŸ”’

upsert_course_research_form ğŸ”’

PUT /api/v0/cms/courses/:course_id/research-consent-form - Upserts courses research form from Gutenberg research form edit.

upsert_course_research_form_questions ğŸ”’

PUT /api/v0/cms/courses/:course_id/research-consent-form-questions - Upserts questions for the courses research form from Gutenberg research form edit.