headless_lms_server::controllers::cms

Module gutenberg

Structs

ThemeResponse

ThemeSupports

Statics

APP_USER_AGENT 🔒

Functions

_add_routes

Add a route for each controller in this module.

get_mentimeter_oembed_data 🔒

get_oembed_data_from_provider 🔒

GET /api/v0/cms/gutenberg/oembed/preview?url=https%3A%2F%2Fyoutube.com%2Fwatch%3Fv%3D123123123 - Fetch oembed response from correct provider. Endpoint for proxying oembed requests to correct provider using url query param.

get_theme_settings 🔒

GET /api/v0/cms/gutenberg/themes?context=edit&status=active&_locale=user - Mock themes response Endpoint for proxying themes requests. https://github.com/WordPress/gutenberg/blob/trunk/packages/block-library/src/embed/test/index.native.js#L128

get_thinglink_oembed_data 🔒

get_vimeo_oembed_data 🔒