Function headless_lms_server::controllers::cms::organizations::add_media

async fn add_media(
    organization_id: Path<Uuid>, 
    payload: Multipart, 
    request: HttpRequest, 
    pool: Data<PgPool>, 
    user: AuthUser, 
    file_store: Data<dyn FileStore>, 
    app_conf: Data<ApplicationConfiguration>
) -> ControllerResult<Json<UploadResult>>

POST /api/v0/cms/organizations/:organization_id/upload - Uploads a media (image, audio, file) for the course from Gutenberg page edit.

Put the the contents of the media in a form and add a content type header multipart/form-data.

Example

Request:

POST /api/v0/cms/pages/d86cf910-4d26-40e9-8c9c-1cc35294fdbb/upload HTTP/1.1
Content-Type: multipart/form-data

BINARY_DATA

Response TypeScript definition

type UploadResult = { url: string }

Example response

{
  "url": "http://project-331.local/api/v0/files/courses/1b89e57e-8b57-42f2-9fed-c7a6736e3eec/courses/d86cf910-4d26-40e9-8c9c-1cc35294fdbb/images/iHZMHdvsazy43ZtP0Ea01sy8AOpUiZ.png"
}