headless_lms_server::controllers::course_material

Module chapters

Controllers for requests starting with /api/v0/course_material/chapters.

Functions

_add_routes

Add a route for each controller in this module.

get_chapter_lock_preview 🔒

GET /api/v0/course-material/chapters/:chapter_id/lock-preview - Preview lock chapter

get_chapters_exercises 🔒

GET /api/v0/course-material/chapters/:chapter_id/exercises - Returns a list of pages and its exercises in chapter.

get_chapters_pages_without_main_frontpage 🔒

GET /api/v0/course-material/chapters/:chapter_id/pages-exclude-mainfrontpage - Returns a list of pages in chapter mainfrontpage excluded.

get_public_chapter_pages 🔒

GET /api/v0/course-material/chapters/:chapter_id/pages - Returns a list of pages in chapter.

lock_chapter 🔒

POST /api/v0/course-material/chapters/:chapter_id/lock - Complete chapter (mark as done)