headless_lms_server::controllers::course_material
Module course_instances

Controllers for requests starting with /api/v0/course-material/course-instances.

Structs

SaveCourseSettingsPayload

Functions

_add_routes

get_all_get_all_course_module_completions_for_user_by_course_instance_id 🔒

GET /course-instances/:id/course-module-completions/:user_id - Returns a list of all course module completions for a given user for this course instance.

get_background_questions_and_answers 🔒

GET /api/v0/course-material/course-instance/:course_instance_id/background-questions-and-answers - Gets background questions and answers for an course instance.

get_module_completions_for_course_instance 🔒

GET /api/v0/course-material/course-instance/{course_instance_id}/module-completions

get_user_progress_for_course_instance 🔒

GET /api/v0/course-material/course-instance/:course_intance_id/progress - returns user progress information.

get_user_progress_for_course_instance_chapter 🔒

GET `/api/v0/course-material/course-instance/:course_instance_id/chapters/:chapter_id/progress - Returns user progress for chapter in course instance.

get_user_progress_for_course_instance_chapter_exercises 🔒

GET /api/v0/course-material/course-instance/:course_instance_id/chapters/:chapter_id/exercises/progress - Returns user progress for an exercise in given course instance.

save_course_settings 🔒

POST /api/v0/course-material/course-instance/:course_instance_id/save-course-settings - enrolls user to the course instance and save background questions.