Module headless_lms_server::controllers::course_material::courses

Controllers for requests starting with /api/v0/course-material/courses.

Structs

ChaptersWithStatus

CourseMaterialCourseModule

RequestInformation ğŸ”’

Functions

_add_routes

Add a route for each controller in this module.

collect_course_modules ğŸ”’

Combines course modules and chapters, consuming them.

derive_information_from_requester ğŸ”’

Used in get_course_page_by_path for path for anonymous visitor counts

feedback

POST /api/v0/course-material/courses/:course_id/feedback - Creates new feedback.

get_chapters ğŸ”’

GET /api/v0/course-material/courses/:course_id/chapters - Returns a list of chapters in a course.

get_course ğŸ”’

GET /api/v0/course-material/courses/:course_id - Get course.

get_course_instances ğŸ”’

GET /api/v0/course-material/courses/:course_id/course-instances - Returns all course instances for given course id.

get_course_page_by_path ğŸ”’

GET /:course_slug/page-by-path/... - Returns a course page by path

get_current_course_instance ğŸ”’

GET /api/v0/course-material/courses/:course_id/current-instance - Returns the instance of a course for the current user, if there is one.

get_material_references_by_course_id ğŸ”’

Response TypeScript definition

get_public_course_pages ğŸ”’

GET /api/v0/course-material/courses/:course_id/pages - Returns a list of public pages on a course.

get_public_top_level_pages ğŸ”’

GET /api/v0/course-material/courses/:course_id/top-level-pages

get_user_course_settings ğŸ”’

GET /api/v0/course-material/courses/:course_id/user-settings - Returns user settings for the current course.

glossary ğŸ”’

Response TypeScript definition

propose_edit ğŸ”’

POST /api/v0/course-material/courses/:course_slug/edit - Creates a new edit proposal.

search_pages_with_phrase ğŸ”’

POST /api/v0/course-material/courses/:course_id/search-pages-with-phrase - Returns a list of pages given a search query.

search_pages_with_words ğŸ”’

POST /api/v0/course-material/courses/:course_id/search-pages-with-words - Returns a list of pages given a search query.