Function headless_lms_server::controllers::course_material::courses::search_pages_with_words

async fn search_pages_with_words(
    course_id: Path<Uuid>,
    payload: Json<SearchRequest>,
    pool: Data<PgPool>
) -> ControllerResult<Json<Vec<PageSearchResult>>>

POST /api/v0/course-material/courses/:course_id/search-pages-with-words - Returns a list of pages given a search query.

Provided words can appear in any order in the source document.

Example

Request:

POST /api/v0/course-material/courses/1a68e8b0-d151-4c0e-9307-bb154e9d2be1/search-pages-with-words HTTP/1.1
Content-Type: application/json

{
  "query": "Everything"
}

Response TypeScript definition

type Vec<PageSearchResult> = Array<{
  id: string
  title_headline: string | null
  rank: number | null
  content_headline: string | null
  url_path: string
}>

Example response

[
  {
    "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "title_headline": "Introduction to everything",
    "rank": 0.6079271,
    "content_headline": null,
    "url_path": "/chapter-1"
  },
  {
    "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "title_headline": "In the second chapter...",
    "rank": 0.24317084,
    "content_headline": "<b>Everything</b> is a big topic",
    "url_path": "/chapter-2"
  }
]