headless_lms_server::controllers::course_material::exercises
Function get_peer_review_for_exercise

async fn get_peer_review_for_exercise(
    pool: Data<PgPool>,
    exercise_id: Path<Uuid>,
    user: AuthUser,
    jwt_key: Data<JwtKey>,
) -> ControllerResult<Json<CourseMaterialPeerOrSelfReviewDataWithToken>>

GET /api/v0/course-material/exercises/:exercise_id/peer-review - Get peer review for an exercise. This includes the submission to peer review and the questions the user is supposed to answer.ALTER

This request will fail if the user is not in the peer review stage yet because the information included in the peer review often exposes the correct solution to the exercise.