headless_lms_server::controllers
Module files

Handlers for HTTP requests to /api/v0/files.

Re-exports

pub use crate::domain::authorization::AuthorizationToken;

pub use crate::domain::models_requests::UploadClaim;

Functions

_add_routes

Add a route for each controller in this module.

redirect_to_storage_service 🔒

GET /api/v0/files/\* Redirects the request to a file storage service.

serve_upload 🔒

GET /api/v0/files/uploads/\* Serve local uploaded file, mostly for development.

upload_from_exercise_service 🔒

POST /api/v0/files/:exercise_service_slug Used to upload data from exercise service iframes.