Function headless_lms_server::controllers::files::upload_from_exercise_service

async fn upload_from_exercise_service(
    pool: Data<PgPool>,
    exercise_service_slug: Path<String>,
    payload: Multipart,
    file_store: Data<dyn FileStore>,
    user: Option<AuthUser>,
    upload_claim: Result<UploadClaim<'static>, ControllerError>,
    app_conf: Data<ApplicationConfiguration>
) -> ControllerResult<Json<HashMap<String, String>>>

POST /api/v0/files/:exercise_service_slug Used to upload data from exercise service iframes.

Returns

The randomly generated paths to each uploaded file in a file_name => file_path hash map.

Response TypeScript definition

type HashMap<String, String> = Record<string, string>

Example response

{
  "key1": "val1",
  "key2": "val2"
}