headless_lms_server::controllers

Module langs

Handlers for HTTP requests to /api/v0/langs. Contains endpoints for the use of tmc-langs.

Structs

SubmissionForm 🔒

Functions

_add_routes

get_course_instance_exercises 🔒

GET /api/v0/langs/course-instances/:id/exercisesReturns the user’s exercise slides for the given course instance.Does not return anything for chapters which are not open yet.Selects slides for exercises with no slide selected yet.Only returns slides which have tasks that are compatible with langs.

get_course_instances 🔒

GET /api/v0/langs/course-instancesReturns the course instances that the user is currently enrolled on that contain TMC exercises.

get_exercise 🔒

GET /api/v0/langs/exercises/:idReturns an exercise slide for the user for the given exercise.Only returns slides

get_submission_grading 🔒

submit_exercise 🔒

POST /api/v0/langs/exercises/:id/submit