Module headless_lms_server::controllers::main_frontend::exams

Structs

ExamCourseInfo

Functions

_add_routes

Add a route for each controller in this module.

duplicate_exam 🔒

POST /api/v0/cms/exams/:exam_id/duplicate - duplicates existing exam.

edit_exam 🔒

POST /api/v0/main-frontend/organizations/{organization_id}/edit-exam - edits an exam.

export_points

GET /api/v0/main-frontend/exams/:id/export-points

export_submissions

GET /api/v0/main-frontend/exams/:id/export-submissions

get_exam

GET `/api/v0/main-frontend/exams/:id

get_exercise_slide_submissions_and_user_exercise_states_with_exam_id 🔒

GET /api/v0/main-frontend/exam/:exam_id/submissions-with-exam-id - Returns all the exercise submissions and user exercise states with exam_id.

get_exercise_slide_submissions_and_user_exercise_states_with_exercise_id 🔒

GET /api/v0/main-frontend/exam/:exercise_id/submissions-with-exercise_id - Returns all the exercise submissions and user exercise states with exercise_id.

get_exercises_with_exam_id 🔒

GET /api/v0/main-frontend/exam/:exam_id/exam-exercises - Returns all the exercises with exam_id.

release_grades 🔒

POST /api/v0/main-frontend/exam/:exam_id/release-grades - Publishes grading results of an exam by updating user_exercise_states according to teacher_grading_decisons and changes teacher_grading_decisions hidden field to false.

set_course

POST /api/v0/main-frontend/exams/:id/set

unset_course

POST /api/v0/main-frontend/exams/:id/unset