Module headless_lms_server::controllers::main_frontend::exercise_slide_submissions

Structs

ExerciseStateIds

Functions

_add_routes

add_teacher_grading 🔒

PUT /api/v0/main-frontend/exercise-slide-submissions/add_teacher_grading" - Adds a new teacher grading decision, without updating user exercise state

get_submission_info 🔒

GET /api/v0/main-frontend/exercise-slide-submissions/{submission_id}/info"- Returns data necessary for rendering a submission.

get_user_exercise_state_info 🔒

GET /api/v0/main-frontend/exercise-slide-submissions/{exam_id}/{exercise_id}/{user_id}/user-exercise-state-info-

update_answer_requiring_attention 🔒

PUT /api/v0/main-frontend/exercise-slide-submissions/update-answer-requiring-attention" - Given a teacher grading decision, updates an answer by giving it a manual score given.