Module headless_lms_server::controllers::main_frontend::regradings

Controllers for requests starting with /api/v0/main-frontend/regradings/.

Functions

_add_routes

Add a route for each controller in this module.

create_by_exercise_task_submission_ids 🔒

POST /api/v0/main-frontend/regradings - Creates a new regrading for the supplied exercise task submission ids and returns the new regrading id.

get_regrading_info_by_id 🔒

GET /api/v0/main-frontend/regradings/{id} - Returns relevant information about a regrading.

get_regradings 🔒

GET /api/v0/main-frontend/regradings - Returns a paginated list of all the regradings.

get_regradings_count 🔒

GET /api/v0/main-frontend/regradings/count - Counts regradings