Module headless_lms_server::controllers::main_frontend::user_details

Functions

_add_routes

get_user_details

GET /api/v0/main-frontend/user-details/[id] - Find user details by user id

search_users_by_email

GET /api/v0/main-frontend/user-details/search-by-email - Allows to search user by their email

search_users_by_other_details

GET /api/v0/main-frontend/user-details/search-by-other-details - Allows to search user by their names etc.

search_users_fuzzy_match

GET /api/v0/main-frontend/user-details/search-fuzzy-match - Allows to find the right user details in cases where there is a small typing error in the search query