Module headless_lms_server::controllers::study_registry

Handlers for HTTP requests to /api/v0/study-registry.

The study registry provides an access to student completion records. It is generally only available to authorized study registries, meaning that most endpoints will require a valid authorization token to access.

When accessing study registry, the authorization token should be given as the following header:

Authorization: Basic documentationOnlyExampleSecretKey-12345

For more details, please view the submodules.

Modules

completion_registered_to_study_registry

Controllers for requests starting with /api/v0/study-registry/completion-registered-to-study-registry

completions

Controllers for requests starting with /api/v0/study-registry/completions