headless_lms_server::controllers::study_registry

Module completions

Controllers for requests starting with /api/v0/study-registry/completions

The study registry provides an access to student completion records. It is generally only available to authorized study registries, meaning that most endpoints will require a valid authorization token to access.

When accessing study registry, the authorization token should be given as the following header:

Authorization: Basic documentationOnlyExampleSecretKey-12345

For more details, please view the individual functions.

Structs

GetCompletionsQueryParamers 🔒

Functions

get_completions 🔒

GET /api/v0/study-registry/completions/[:course_id | :uh_course_code | :course_slug] – Get completions from all modules in a course.

get_module_completions 🔒

GET /api/v0/study-registry/completions/[:course_id | :uh_course_code | :course_slug]/:course_module_id – Get completions from a single course module.