Module headless_lms_server::controllers::tmc_server::users_by_upstream_id

Handlers for HTTP requests to /api/v0/tmc-server/users-by-upstream-id.

These endpoints are used by the TMC server so that it can integrate with this system.

Functions

_add_routes

get_user_by_upstream_id

GET /api/v0/tmc-server/users-by-upstream-id/:id Endpoint that TMC server uses to get user information by using its own ids.