headless_lms_server

Module domain

Helpful structs and functions that are not related to database tables.

Modules

authorization

Common functionality related to authorization

csv_export

error

Contains error and result types for all the controllers.

exercise_repositories

exercises

langs

models_requests

Contains helper functions that are passed to headless-lms-models where it needs to make requests to exercise services.

oauth

rate_limit_middleware_builder

request_id

request_span_middleware

Middleware that wraps HTTP requests to tokio tracing spans for debugging and attaches a request id to all log messages.

system_health

System health checking logic for Kubernetes resources.