headless_lms_server::domain::authorization
Function verify_auth_user_exists

async fn verify_auth_user_exists(
    auth_user: AuthUser,
    pool: Option<&Data<PgPool>>,
    session: &Session,
) -> Result<AuthUser, ControllerError>

For making sure the user saved in the session still exists in the database. Check the user’s existance when the session is at least 3 hours old, updates the session automatically, and returns an up-to-date AuthUser.