Module headless_lms_server::domain::chatbot

Structs

ApiChatMessage

The format accepted by the api.

ChatReponse

ChatRequest

Choice

ContentFilter

ContentFilterResults

Delta

RequestCancelledGuard 🔒

ResponseChunk

Constants

CHATBOT_AZURE_API_VERSION

Statics

CLIENT 🔒

Functions

estimate_tokens

Estimate the number of tokens in a given text. We use this for example to estimate the expense of a chat request. The result is not accurate but it is cheap to calculate.

send_chat_request_and_parse_stream