Function headless_lms_server::domain::csv_export::course_module_completion_info_to_grade_string

fn course_module_completion_info_to_grade_string(
    input: Option<&CourseModuleCompletionWithRegistrationInfo>
) -> String

For csv export. Return the grade as a number if there is a numeric grade. If the grade is not numeric, returns pass/fail/ If course module has not been completed yet, returns “-”.