Struct headless_lms_server::domain::csv_export::CsvWriter

struct CsvWriter<W: Write> {
    csv_writer: Arc<Mutex<Writer<W>>>,
    handles: FuturesUnordered<JoinHandle<Result<()>>>,
}

Convenience struct for creating CSV data.

Fields
csv_writer: Arc<Mutex<Writer<W>>>handles: FuturesUnordered<JoinHandle<Result<()>>>
Implementations

impl<W: Write + Send + 'static> CsvWriter<W>

async fn new_with_initialized_headers<I, T>(
writer: W,
headers: I
) -> Result<Self>where
I: IntoIterator<Item = T> + Send + 'static,
T: AsRef<[u8]>,

Creates a new CsvWriter, and also writes the given headers before returning.

fn write_record<I, T>(&self, csv_row: I)where
I: IntoIterator<Item = T> + Send + 'static,
T: AsRef<[u8]>,

Spawns a task that writes a single CSV record

async fn finish(self) -> Result<W>

Waits for handles to finish, flushes the writer and extracts the inner writer. Should always be called before dropping the writer to make sure writing the CSV finishes properly.