headless_lms_server::domain

Module error

Contains error and result types for all the controllers.

Structs

ControllerError

Represents error messages that are sent in responses. Used as the error type in ControllerError, which is used by all the controllers in the application.

ErrorResponse

The format all error messages from the API is in

Enums

ControllerErrorType

The type of ControllerError that occured.

ErrorData

Type Aliases

ControllerResult

Used as the result types for all controllers. Only put information here that you want to be visible to users.