Type Alias headless_lms_server::domain::error::ControllerResult

pub type ControllerResult<T, E = ControllerError> = Result<AuthorizedResponse<T>, E>;

Used as the result types for all controllers. Only put information here that you want to be visible to users.

See also ControllerError for documentation on how to return errors from controllers.

Aliased Type

enum ControllerResult<T, E = ControllerError> {
    Ok(AuthorizedResponse<T>),
    Err(E),
}

Variants

Ok(AuthorizedResponse<T>)

Contains the success value

Err(E)

Contains the error value