headless_lms_server::domain

Module exercise_repositories

Structs

FoundExercise 🔒

InnerArgs 🔒

StoredRepositoryExercise

Functions

contains_tmcignore 🔒

create_and_upload_exercise 🔒

delete

Marks the exercises and repository as deleted and removes the associated files from the file store. Only returns the last error if there are multiple errors when trying to remove the files.

find_exercise_directories 🔒

is_hidden_dir 🔒

is_in_git_dir 🔒

process

Processes an exercise repository, creating a repository exercise for each exercise in it. Each exercise is compressed and uploaded to file storage.

process_inner 🔒

update

Updates the given repository using the given url. Exercises with a known checksum but changed part or name are updated to reflect the new part or name. Exercises with a known part and name but changed checksum are updated in the file store and the checksum updated. Errors may leave some exercises updated and others not, since there’s no mechanism for rolling back any file store updates. However, these inconsistencies will be fixed after a successful retry.

update_exercise 🔒

update_inner 🔒