headless_lms_server::domain::exercise_repositories
Function update

pub async fn update(
    conn: &mut PgConnection,
    repository: Uuid,
    url: &str,
    file_store: &dyn FileStore,
    app_conf: &ApplicationConfiguration,
) -> Result<()>

Updates the given repository using the given url. Exercises with a known checksum but changed part or name are updated to reflect the new part or name. Exercises with a known part and name but changed checksum are updated in the file store and the checksum updated. Errors may leave some exercises updated and others not, since there’s no mechanism for rolling back any file store updates. However, these inconsistencies will be fixed after a successful retry.