Function headless_lms_server::domain::models_requests::make_spec_fetcher

pub fn make_spec_fetcher(
    jwt_key: Arc<JwtKey>
) -> impl Fn(Url, &str, Option<&Value>) -> BoxFuture<'static, ModelResult<Value>>

Fetches a public/model spec based on the private spec from the given url. The slug and jwt key are used for an upload claim that allows the service to upload files as part of the spec.