Trait headless_lms_server::prelude::FileStore

pub trait FileStore {
    fn upload<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        path: &'life1 Path,
        contents: Vec<u8, Global>,
        mime_type: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<(), UtilError>> + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn upload_stream<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        path: &'life1 Path,
        contents: Pin<Box<dyn Stream<Item = Result<Bytes, Error>> + 'static, Global>>,
        mime_type: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<(), UtilError>> + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn download<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 Path
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8, Global>, UtilError>> + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn download_stream<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 Path
    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn Stream<Item = Result<Bytes, Error>> + 'static, Global>, UtilError>> + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_direct_download_url<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 Path
    ) -> Pin<Box<dyn Future<Output = Result<String, UtilError>> + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn delete<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 Path
    ) -> Pin<Box<dyn Future<Output = Result<(), UtilError>> + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn get_download_url(
        &self,
        path: &Path,
        app_conf: &ApplicationConfiguration
    ) -> String { ... }
}

Allows storing files to a file storage backend.

Required Methods

fn upload<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
path: &'life1 Path,
contents: Vec<u8, Global>,
mime_type: &'life2 str
) -> Pin<Box<dyn Future<Output = Result<(), UtilError>> + 'async_trait, Global>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Upload a file that’s in memory to a path.

fn upload_stream<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
path: &'life1 Path,
contents: Pin<Box<dyn Stream<Item = Result<Bytes, Error>> + 'static, Global>>,
mime_type: &'life2 str
) -> Pin<Box<dyn Future<Output = Result<(), UtilError>> + 'async_trait, Global>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Upload a file without loading the whole file to memory

fn download<'life0, 'life1, 'async_trait>(
&'life0 self,
path: &'life1 Path
) -> Pin<Box<dyn Future<Output = Result<Vec<u8, Global>, UtilError>> + 'async_trait, Global>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Download a file to memory.

fn download_stream<'life0, 'life1, 'async_trait>(
&'life0 self,
path: &'life1 Path
) -> Pin<Box<dyn Future<Output = Result<Box<dyn Stream<Item = Result<Bytes, Error>> + 'static, Global>, UtilError>> + 'async_trait, Global>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Download a file without loading the whole file to memory.

fn get_direct_download_url<'life0, 'life1, 'async_trait>(
&'life0 self,
path: &'life1 Path
) -> Pin<Box<dyn Future<Output = Result<String, UtilError>> + 'async_trait, Global>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get a url that can be used to download the file without authentication for a while. In most cases you probably want to use get_download_url() instead.

fn delete<'life0, 'life1, 'async_trait>(
&'life0 self,
path: &'life1 Path
) -> Pin<Box<dyn Future<Output = Result<(), UtilError>> + 'async_trait, Global>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Delete a file.

Provided Methods

fn get_download_url(
&self,
path: &Path,
app_conf: &ApplicationConfiguration
) -> String

Get a url for a file in FileStore that can be used to access the resource.