headless_lms_server
Module programs

Executable programs that can be started. Contains for example the server program, background services, and utility programs.

Modules

calculate_page_visit_stats

chatbot_syncer

doc_file_generator

The doc file generator is used to write example JSON and TypeScript definitions for the docs of return values of API endpoints.

email_deliver

ended_exams_processor

mailchimp_syncer

open_university_registration_link_fetcher

peer_review_updater

regrader

seed

service_info_fetcher

sorter

Manually sorts…

start_server

sync_tmc_users

Syncs tmc users