headless_lms_server::programs

Module open_university_registration_link_fetcher

Structs

OpenUniversityInfo 🔒

Constants

OPEN_UNIVERSITY_COURSE_URL 🔒

OPEN_UNIVERSITY_REGISTRATION_BASE_URL 🔒

OPEN_UNIVERSITY_TOKEN 🔒

Functions

fetch_and_update_completion_links 🔒

Fetches up-to-date Open University completion registration links, upserts them to database and returns to amount of updated records.

get_open_university_info_for_course_code 🔒

main

select_best_candidate 🔒

update_course_registration_link 🔒