headless_lms_server::programs::open_university_registration_link_fetcher
Function fetch_and_update_completion_links

async fn fetch_and_update_completion_links(
    conn: &mut PgConnection,
    open_university_course_url: &str,
    open_university_token: &str,
) -> Result<u32>

Fetches up-to-date Open University completion registration links, upserts them to database and returns to amount of updated records.