Crate headless_lms_server

The server that handles the requests.

See also

headless_lms_utils

headless_lms_models

Modules

controllers

Handlers for HTTP requests to /api/v0.

domain

Helpful structs and functions that are not related to database tables.

prelude

Re-exports commonly used types for convenient use across the crate. Intended to be glob-imported like use crate::prelude::*;.

programs

Executable programs that can be started. Contains for example the server program, background services, and utility programs.

Macros

generated_docs

Includes the type’s JSON example generated by doc-file-generator as a string. Used with the helper macro from the doc-macro crate: #generated_doc

Functions

configure

setup_tracing

Sets up tokio tracing. Also makes sure that log statements from libraries respect the log level settings that have been set with RUST_LOG, for example:

Type Definitions

OAuthClient