Module headless_lms_server::controllers::auth

Handlers for HTTP requests to /api/v0/auth.

Structs

CreateAccountDetails

GraphQLRquest

Login

MoocfiCurrentUser

MoocfiCurrentUserResponse

MoocfiCurrentUserResponseData

Functions

_add_routes

async_http_client_with_headers 🔒

HTTP Client used only for authing with TMC server, this is to ensure that TMC server does not rate limit auth requests from backend

authorize_action_on_resource

POST /api/v0/auth/authorize checks whether user can perform specified action on specified resource.

authorize_multiple_actions_on_resources

POST /api/v0/auth/authorize-multiple checks whether user can perform specified action on specified resource. Returns booleans for the authorizations in the same order as the input.

get_user_from_moocfi

logged_in

GET /api/v0/auth/logged-in Returns the current user’s login status.

login

POST /api/v0/auth/login Logs in to TMC.

logout

POST /api/v0/auth/logout Logs out.

post_new_user_to_moocfi

Posts new user account to tmc.mooc.fi.

signup

POST /api/v0/auth/signup Creates new mooc.fi account and signs in.

Type Definitions

LoginToken