Function headless_lms_server::controllers::auth::authorize_multiple_actions_on_resources

pub async fn authorize_multiple_actions_on_resources(
    pool: Data<PgPool>,
    user: Option<AuthUser>,
    payload: Json<Vec<ActionOnResource>>,
) -> ControllerResult<Json<Vec<bool>>>

POST /api/v0/auth/authorize-multiple checks whether user can perform specified action on specified resource. Returns booleans for the authorizations in the same order as the input.