Function headless_lms_server::controllers::course_material::courses::get_course_page_by_path

async fn get_course_page_by_path(
    params: Path<(String, String)>, 
    pool: Data<PgPool>, 
    user: Option<AuthUser>, 
    ip_to_country_mapper: Data<IpToCountryMapper>, 
    req: HttpRequest
) -> ControllerResult<Json<CoursePageWithUserData>>

GET /:course_slug/page-by-path/... - Returns a course page by path

If the page has moved and there’s a redirection, this will still return the moved page but the field was_redirected will indicate that the redirection happened. The new path can be found in the page object. The frontend is supposed to update the url of the page to the new location without reloading the page.

Example

GET /api/v0/course-material/courses/introduction-to-everything/page-by-path//part-2/hello-world

Response TypeScript definition

type CoursePageWithUserData = {
  page: Page
  instance: CourseInstance | null
  settings: UserCourseSettings | null
  was_redirected: boolean
  is_test_mode: boolean
}

Example response

{
  "page": {
    "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "created_at": "2021-12-31T22:00:00Z",
    "updated_at": "2021-12-31T22:00:00Z",
    "course_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "exam_id": null,
    "chapter_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "url_path": "/part-1/hello-world",
    "title": "Hello world!",
    "deleted_at": null,
    "content": [
      {
        "attributes": {
          "hidden": false
        },
        "clientId": "c68f55ae-65c4-4e9b-aded-0b52e36e344a",
        "innerBlocks": [],
        "isValid": true,
        "name": "moocfi/pages-in-part"
      },
      {
        "attributes": {
          "hidden": false
        },
        "clientId": "415ecc4c-a5c6-410e-a43f-c14b8ee910ea",
        "innerBlocks": [],
        "isValid": true,
        "name": "moocfi/exercises-in-part"
      }
    ],
    "order_number": 123,
    "copied_from": null
  },
  "instance": {
    "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "created_at": "2021-12-31T22:00:00Z",
    "updated_at": "2021-12-31T22:00:00Z",
    "deleted_at": null,
    "course_id": "3c6ca496-17ac-445c-88c0-4ded2f2dbe58",
    "starts_at": "2021-12-31T22:00:00Z",
    "ends_at": null,
    "name": "Instance",
    "description": "Description",
    "teacher_in_charge_name": "Example Teacher",
    "teacher_in_charge_email": "example@email.com",
    "support_email": "example@email.com"
  },
  "settings": {
    "user_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "course_language_group_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "created_at": "2021-12-31T22:00:00Z",
    "updated_at": "2021-12-31T22:00:00Z",
    "deleted_at": null,
    "current_course_id": "3c6ca496-17ac-445c-88c0-4ded2f2dbe58",
    "current_course_instance_id": "307fa56f-9853-4f5c-afb9-a6736c232f32"
  },
  "was_redirected": false,
  "is_test_mode": false
}