Function headless_lms_server::controllers::course_material::exercises::get_exercise

async fn get_exercise(
    pool: Data<PgPool>,
    exercise_id: Path<Uuid>,
    user: Option<AuthUser>
) -> ControllerResult<Json<CourseMaterialExercise>>

GET /api/v0/course-material/exercises/:exercise_id - Get exercise by id. Includes relevant context so that doing the exercise is possible based on the response.

This endpoint does not expose exercise’s private spec because it would expose the correct answers to the user.

Response TypeScript definition

type CourseMaterialExercise = {
  exercise: Exercise
  can_post_submission: boolean
  current_exercise_slide: CourseMaterialExerciseSlide
  exercise_status: ExerciseStatus | null
  exercise_slide_submission_counts: Record<string, number>
  peer_review_config: CourseMaterialPeerReviewConfig | null
  previous_exercise_slide_submission: ExerciseSlideSubmission | null
}

Example response

{
  "exercise": {
    "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "created_at": "2021-12-31T22:00:00Z",
    "updated_at": "2021-12-31T22:00:00Z",
    "name": "Hello Exercise",
    "course_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "exam_id": null,
    "page_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "chapter_id": null,
    "deadline": null,
    "deleted_at": null,
    "score_maximum": 1,
    "order_number": 123,
    "copied_from": null,
    "max_tries_per_slide": 17,
    "limit_number_of_tries": true,
    "needs_peer_review": false,
    "use_course_default_peer_review_config": false
  },
  "can_post_submission": true,
  "current_exercise_slide": {
    "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "exercise_tasks": [
      {
        "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
        "exercise_service_slug": "example-exercise",
        "exercise_slide_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
        "exercise_iframe_url": "http://project-331.local/example-exercise/exercise",
        "pseudonumous_user_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
        "assignment": {
          "attributes": {
            "content": "Answer this question.",
            "dropCap": false
          },
          "clientId": "187a0aea-c088-4354-a1ea-f0cab082c065",
          "innerBlocks": [],
          "isValid": true,
          "name": "core/paragraph"
        },
        "public_spec": [
          {
            "id": "7ab2591c-b0f3-4543-9548-a113849b0f94",
            "name": "a"
          },
          {
            "id": "a833d1df-f27b-4fbf-b516-883a62c09d88",
            "name": "b"
          },
          {
            "id": "03d4b3d4-88af-4125-88b7-4ee052fd876f",
            "name": "c"
          }
        ],
        "model_solution_spec": null,
        "previous_submission": {
          "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
          "created_at": "2021-12-31T22:00:00Z",
          "updated_at": "2021-12-31T22:00:00Z",
          "deleted_at": null,
          "exercise_slide_submission_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
          "exercise_task_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
          "exercise_slide_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
          "data_json": {
            "choice": "a"
          },
          "exercise_task_grading_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
          "metadata": null
        },
        "previous_submission_grading": {
          "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
          "created_at": "2021-12-31T22:00:00Z",
          "updated_at": "2021-12-31T22:00:00Z",
          "exercise_task_submission_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
          "course_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
          "exam_id": null,
          "exercise_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
          "exercise_task_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
          "grading_priority": 1,
          "score_given": 80.0,
          "grading_progress": "FullyGraded",
          "unscaled_score_given": 80.0,
          "unscaled_score_maximum": 100,
          "grading_started_at": "2021-12-31T22:00:00Z",
          "grading_completed_at": "2021-12-31T22:00:00Z",
          "feedback_json": null,
          "feedback_text": null,
          "deleted_at": null
        },
        "order_number": 1
      }
    ]
  },
  "exercise_status": {
    "score_given": null,
    "activity_progress": "InProgress",
    "grading_progress": "NotReady",
    "reviewing_stage": "NotStarted"
  },
  "exercise_slide_submission_counts": {
    "7dea54af-3d38-4f7c-8969-ecb17b55ec02": 4,
    "2794a98e-d594-40cf-949e-7cc011755a58": 2
  },
  "peer_review_config": {
    "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "course_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "exercise_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "peer_reviews_to_give": 3,
    "peer_reviews_to_receive": 2
  },
  "previous_exercise_slide_submission": {
    "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "created_at": "2021-12-31T22:00:00Z",
    "updated_at": "2021-12-31T22:00:00Z",
    "deleted_at": null,
    "exercise_slide_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "course_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "course_instance_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "exam_id": null,
    "exercise_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "user_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "user_points_update_strategy": "CanAddPointsAndCanRemovePoints"
  }
}