Function headless_lms_server::controllers::course_material::exercises::get_peer_review_for_exercise

async fn get_peer_review_for_exercise(
    pool: Data<PgPool>,
    exercise_id: Path<Uuid>,
    user: AuthUser,
    jwt_key: Data<JwtKey>
) -> ControllerResult<Json<CourseMaterialPeerReviewDataWithToken>>

GET /api/v0/course-material/exercises/:exercise_id/peer-review - Get peer review for an exercise. This includes the submission to peer review and the questions the user is supposed to answer.ALTER

This request will fail if the user is not in the peer review stage yet because the information included in the peer review often exposes the correct solution to the exercise.

Response TypeScript definition

type CourseMaterialPeerReviewDataWithToken = {
  course_material_peer_review_data: CourseMaterialPeerReviewData
  token: string | null
}

Example response

{
  "course_material_peer_review_data": {
    "answer_to_review": {
      "exercise_slide_submission_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
      "course_material_exercise_tasks": [
        {
          "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
          "exercise_service_slug": "example-exercise",
          "exercise_slide_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
          "exercise_iframe_url": "http://project-331.local/example-exercise/exercise",
          "pseudonumous_user_id": "934ac548-f7de-479a-b4f8-af9c0c6c22dc",
          "assignment": {
            "attributes": {
              "content": "Answer this question.",
              "dropCap": false
            },
            "clientId": "187a0aea-c088-4354-a1ea-f0cab082c065",
            "innerBlocks": [],
            "isValid": true,
            "name": "core/paragraph"
          },
          "public_spec": [
            {
              "id": "7ab2591c-b0f3-4543-9548-a113849b0f94",
              "name": "a"
            },
            {
              "id": "a833d1df-f27b-4fbf-b516-883a62c09d88",
              "name": "b"
            },
            {
              "id": "03d4b3d4-88af-4125-88b7-4ee052fd876f",
              "name": "c"
            }
          ],
          "model_solution_spec": null,
          "previous_submission": {
            "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
            "created_at": "2021-12-31T22:00:00Z",
            "updated_at": "2021-12-31T22:00:00Z",
            "deleted_at": null,
            "exercise_slide_submission_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
            "exercise_task_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
            "exercise_slide_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
            "data_json": {
              "choice": "a"
            },
            "exercise_task_grading_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
            "metadata": null
          },
          "previous_submission_grading": {
            "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
            "created_at": "2021-12-31T22:00:00Z",
            "updated_at": "2021-12-31T22:00:00Z",
            "exercise_task_submission_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
            "course_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
            "exam_id": null,
            "exercise_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
            "exercise_task_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
            "grading_priority": 1,
            "score_given": 80.0,
            "grading_progress": "PendingManual",
            "unscaled_score_given": 80.0,
            "unscaled_score_maximum": 100,
            "grading_started_at": "2021-12-31T22:00:00Z",
            "grading_completed_at": "2021-12-31T22:00:00Z",
            "feedback_json": null,
            "feedback_text": null,
            "deleted_at": null
          },
          "order_number": 0
        }
      ]
    },
    "peer_review_config": {
      "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
      "created_at": "2021-12-31T22:00:00Z",
      "updated_at": "2021-12-31T22:00:00Z",
      "deleted_at": null,
      "course_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
      "exercise_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
      "peer_reviews_to_give": 3,
      "peer_reviews_to_receive": 2,
      "accepting_threshold": 3.0,
      "accepting_strategy": "AutomaticallyAcceptOrManualReviewByAverage",
      "manual_review_cutoff_in_days": 21
    },
    "peer_review_questions": [
      {
        "id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
        "created_at": "2021-12-31T22:00:00Z",
        "updated_at": "2021-12-31T22:00:00Z",
        "deleted_at": null,
        "peer_review_config_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
        "order_number": 0,
        "question": "Was the answer well thought out?",
        "question_type": "Essay",
        "answer_required": true
      }
    ],
    "num_peer_reviews_given": 2
  },
  "token": "eyJhbGciOiJIUzI1NiJ9.eyJleGVyY2lzZV9zbGlkZV9zdWJtaXNzaW9uX2lkIjoiMzgyNzA0YzMtOTc3Mi00M2NjLTgwMTktMTViMmFjM2QxODI0IiwicGVlcl9yZXZpZXdfY29uZmlnX2lkIjoiYjViZjM1YTctZDdhYS00NGJhLWExODYtYzMwMGFjMTU3MjdhIiwiZXhwaXJhdGlvbl90aW1lIjoiMjAyMy0wMy0yOFQxODo1ODozMC40MTA4NTM3MTdaIn0.jCEgFggGGMaqdzH3p9NMLkZPTG2q-oE7d64glblacfs"
}