Module headless_lms_server::controllers::helpers::file_uploading

Helper functions related to uploading to file storage.

Re-exports

pub use crate::domain::authorization::AuthorizationToken;

Structs

ExerciseTaskInfo

Enums

FileType 🔒

StoreKind

Functions

delete_file_from_storage

generate_audio_path 🔒

Generates a path for an audio file with the appropriate extension.

generate_file_path 🔒

Generates a path for a generic file with the appropriate extension based on its filename.

generate_image_path 🔒

Generates a path for an image file with the appropriate extension.

make_filename_safe 🔒

nested_path 🔒

path 🔒

process_exercise_service_upload

Processes an upload from an exercise service or an exercise iframe. This function assumes that any permission checks have already been made.

upload_certificate_svg

upload_exercise_archive

upload_field_from_cms

Processes an upload from CMS.

upload_field_to_storage 🔒

Uploads the data from the multipart field to the given path in file storage.

upload_file_from_cms

Processes an upload from CMS.

upload_file_to_storage 🔒

upload_image_for_organization

Processes an upload for an organization’s image.

validate_media_headers 🔒

Generates a path for an audio file with the appropriate extension.