headless_lms_server::controllers::main_frontend
Module course_instances

Controllers for requests starting with /api/v0/main-frontend/course-instances.

Functions

_add_routes

Add a route for each controller in this module.

certificate_configurations

GET /course-instances/:id/default-certificate-configurations - gets default certificate configurations of the given course instance. A default certificate configuration requires only one course module to be completed.

completions 🔒

GET /api/v0/main-frontend/course-instances/{course_instance_id}/completions

completions_export

GET /course-instances/:id/export-completions - gets CSV of course completion based on course_instance ID.

delete 🔒

POST /course-instances/:id/delete

edit

POST /course-instances/:id/edit

get_all_exercise_statuses_by_course_instance_id 🔒

GET /course-instances/:id/status-for-all-exercises/:user_id - Returns a status for all exercises in a course instance for a given user.

get_all_get_all_course_module_completions_for_user_by_course_instance_id 🔒

GET /course-instances/:id/course-module-completions/:user_id - Returns a list of all course module completions for a given user for this course instance.

get_course_instance 🔒

GET /course-instances/:id

get_email_templates_by_course_instance_id 🔒

get_user_progress_for_course_instance 🔒

GET /api/v0/main-frontend/course-instance/:course_instance_id/progress/:user_id - returns user progress information.

point_export

GET /api/v0/main-frontend/course-instances/${courseInstanceId}/export-points - gets CSV of course instance points based on course_instance ID.

points 🔒

post_completions 🔒

POST /api/v0/main-frontend/course-instances/{course_instance_id}/completions

post_new_email_template 🔒

post_reprocess_module_completions 🔒

POST /api/v0/main-frontend/course-instances/{course_instance_id}/reprocess-completions

preview_post_completions 🔒