headless_lms_server::controllers::main_frontend
Module exercises

Controllers for requests starting with /api/v0/main-frontend/exercises.

Structs

ExerciseSubmissions

ResetExercisesPayload

Functions

_add_routes

Add a route for each controller in this module.

get_exercise 🔒

GET /api/v0/main-frontend/exercises/:exercise_id - Returns a single exercise.

get_exercise_answers_requiring_attention 🔒

GET /api/v0/main-frontend/exercises/:exercise_id/answers-requiring-attention - Returns an exercise’s answers requiring attention.

get_exercise_submissions 🔒

GET /api/v0/main-frontend/exercises/:exercise_id/submissions - Returns an exercise’s submissions.

get_exercises_by_course_id

GET /api/v0/main-frontend/exercises/:course_id/exercises-by-course-id - Returns all exercises for a course with course_id

reset_exercises_for_selected_users

POST /api/v0/main-frontend/exercises/:course_id/reset-exercises-for-selected-users - Resets all selected exercises for selected users and then logs the resets to exercise_reset_logs table