Module headless_lms_server::controllers::main_frontend::organizations

Controllers for requests starting with /api/v0/main-frontend/organizations.

Functions

_add_routes

Add a route for each controller in this module.

create_exam 🔒

POST /api/v0/main-frontend/organizations/{organization_id}/exams - Creates new exam for the organization.

get_all_organizations 🔒

GET /api/v0/main-frontend/organizations - Returns a list of all organizations.

get_course_exams 🔒

GET /api/v0/main-frontend/organizations/{organization_id}/course_exams - Returns an organizations exams in CourseExam form.

get_org_exams 🔒

GET /api/v0/main-frontend/organizations/{organization_id}/exams - Returns an organizations exams in Exam form.

get_organization 🔒

GET /api/v0/main-frontend/organizations/{organization_id} - Returns an organizations with id.

get_organization_active_courses 🔒

Response TypeScript definition

get_organization_active_courses_count 🔒

Response TypeScript definition

get_organization_course_count 🔒

Response TypeScript definition

get_organization_courses 🔒

GET /api/v0/main-frontend/organizations/{organization_id}/courses" - Returns a list of all courses in a organization.

remove_organization_image 🔒

DELETE /api/v0/main-frontend/organizations/:organizations_id/image - Removes the organizations image.

set_organization_image 🔒

PUT /api/v0/main-frontend/organizations/:organizations_id/image - Sets or updates the chapter image.