Module headless_lms_server::controllers::main_frontend::playground_views

Handles playground-views-related functionality, in particular the websocket connections used to update the grading for services like tmc.

Structs

ClientConnection 🔒

PlaygroundSubmissionMessage 🔒

WsConnections 🔒

Enums

PlaygroundViewsMessage

The message type for all messages sent from the server to the client from the playgrounds-views websocket connection.

Constants

CONNECTION_TIMEOUT 🔒

PING_INTERVAL 🔒

Statics

WS_CONNECTIONS 🔒

Functions

_add_routes

receive_grading 🔒

playground-views passes a URL pointing to this route to an exercise service when sending submissions so that if the service has an update for a pending grading, it will be sent here and passed on to through the websocket

websocket 🔒

Starts a new websocket connection.