Module headless_lms_server::controllers::main_frontend::proposed_edits

Structs

GetEditProposalsQuery

Functions

_add_routes

Add a route for each controller in this module.

get_edit_proposal_count

GET /api/v0/main-frontend/proposed-edits/course/:id/count - Returns the amount of feedback for the given course.

get_edit_proposals

GET /api/v0/main-frontend/proposed-edits/course/:id?pending=true - Returns feedback for the given course.

process_edit_proposal

POST /api/v0/main-frontend/proposed-edits/process-edit-proposal - Processes the given edit proposal.