headless_lms_server::controllers::main_frontend
Module users

Structs

ConsentData

Functions

_add_routes

get_all_research_form_answers_with_user_id 🔒

GET /api/v0/main-frontend/users/get-user-research-consents - Gets all users research consents for a course specific research form.

get_course_instance_enrollments_for_user

GET /api/v0/main-frontend/users/:id/course-instance-enrollments

get_my_courses 🔒

GET /api/v0/main-frontend/users/my-courses - Gets all the courses the user has either started or gotten a permission to.

get_research_consent_by_user_id

GET /api/v0/main-frontend/users/get-user-research-consent - Gets users research consent.

get_user

GET /api/v0/main-frontend/users/:id

get_user_reset_exercise_logs

GET /api/v0/main-frontend/users/:id/user-reset-exercise-logs - Get all logs of reset exercises for a user

post_user_consents

POST /api/v0/main-frontend/users/user-research-consents - Adds a research consent for a student.