Function headless_lms_server::controllers::study_registry::completions::get_module_completions

async fn get_module_completions(
    req: HttpRequest, 
    path: Path<(String, Uuid)>, 
    pool: Data<PgPool>
) -> ControllerResult<HttpResponse>

GET /api/v0/main-frontend/study-registry/completions/:course_id/:course_module_id

GET /api/v0/main-frontend/study-registry/completions/:uh_course_code/:course_module_id

GET /api/v0/main-frontend/study-registry/completions/:course_slug/:course_module_id

Gets all course completions for a submodule of a given course. The course identifier can either be its University of Helsinki course code, or a system-local slug or hash id. For module identifier, only the hash id is supported.

This endpoint is only available to authorized study registries, and requires a valid authorization token to access. Results are also streamed rather than included in the response body. In case of an error during transmission, an error message will be appended to the end of the broken stream output.

Example requests

Using University of Helsinki course code:

GET /api/v0/main-frontend/study-registry/completions/BSCS1001 HTTP/1.1
Authorization: Basic documentationOnlyExampleSecretKey-12345
Content-Type: application/json

Using course slug:

GET /api/v0/main-frontend/study-registry/completions/introduction-to-programming HTTP/1.1
Authorization: Basic documentationOnlyExampleSecretKey-12345
Content-Type: application/json

Using course id:

GET /api/v0/main-frontend/study-registry/completions/b3e9575b-fa13-492c-bd14-10cb27df4eec HTTP/1.1
Authorization: Basic documentationOnlyExampleSecretKey-12345
Content-Type: application/json

Response TypeScript definition

type Vec<StudyRegistryCompletion> = Array<{
  completion_date: Date
  completion_language: string
  completion_registration_attempt_date: Date | null
  email: string
  grade: StudyRegistryGrade
  id: string
  user_upstream_id: string
  tier: number | null
}>

Example response

[
  {
    "completion_date": "2022-06-21T00:00:00Z",
    "completion_language": "en-US",
    "completion_registration_attempt_date": null,
    "email": "student@example.com",
    "grade": {
      "scale": "sis-0-5",
      "grade": "4"
    },
    "id": "633852ce-c82a-4d60-8ab5-28745163f6f9",
    "user_upstream_id": "307fa56f-9853-4f5c-afb9-a6736c232f32",
    "tier": null
  }
]