headless_lms_server::domain
Module models_requests

Contains helper functions that are passed to headless-lms-models where it needs to make requests to exercise services.

Structs

GivePeerReviewClaim

GradingUpdateClaim

JwtKey

SpecRequest

Accepted by the public-spec and model-solution endpoints of exercise services.

UploadClaim

Constants

EXERCISE_SERVICE_GRADING_UPDATE_CLAIM_HEADER 🔒

EXERCISE_SERVICE_UPLOAD_CLAIM_HEADER 🔒

Functions

fetch_service_info

make_grading_request_sender

make_seed_spec_fetcher_with_cache

A caching spec fetcher ONLY FOR THE SEED that returns a cached spec if the same (url, exercise_service_slug, private_spec) is requested. Since this is only used during seeding, there is no cache eviction.

make_spec_fetcher

Fetches a public/model spec based on the private spec from the given url. The slug and jwt key are used for an upload claim that allows the service to upload files as part of the spec.

reqwest_err 🔒

Type Aliases

SpecCache 🔒

A type for caching the spec fetching (only for the seed)