Module headless_lms_server::prelude

Re-exports commonly used types for convenient use across the crate. Intended to be glob-imported like use crate::prelude::*;.

Re-exports

pub use crate::controllers::helpers::file_uploading::upload_file_from_cms;

pub use crate::controllers::helpers::file_uploading::StoreKind;

pub use crate::controllers::UploadResult;

pub use crate::domain::authorization::authorize;

pub use crate::domain::authorization::parse_secret_key_from_header;

pub use crate::domain::authorization::skip_authorize;

pub use crate::domain::authorization::Action as Act;

pub use crate::domain::authorization::AuthUser;

pub use crate::domain::authorization::Resource as Res;

pub use crate::domain::error::ControllerError;

pub use crate::domain::error::ControllerErrorType;

pub use crate::domain::error::ControllerResult;

pub use crate::generated_docs;

pub use headless_lms_models as models;

Modules

web

Essentials helper functions and types for application registration.

Structs

ApplicationConfiguration

HttpRequest

An incoming request.

HttpResponse

An outgoing response.

Multipart

The server-side implementation of multipart/form-data requests.

Pagination

Represents the URL query parameters page and limit, used for paginating database queries.

PgConnection

A connection to a PostgreSQL database.

ServiceConfig

Enables parts of app configuration to be declared separately from the app itself. Helpful for modularizing large applications.

Uuid

A Universally Unique Identifier (UUID).

Enums

PKeyPolicy

Helper struct to use with functions that insert data into the database.

Traits

BackendError

The error types of this program all implement this trait for interoperability.

Connection

Represents a single database connection.

Context

Provides the context method for Result.

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

FileStore

Allows storing files to a file storage backend.

FromRow

A record that can be built from a row returned by the database.

Serialize

A data structure that can be serialized into any data format supported by Serde.

TryToOptional

Type

Indicates that a SQL type is supported for a database.

Type Definitions

PgPool

An alias for Pool, specialized for Postgres.

Derive Macros

Deserialize

FromRow

Serialize

Type