Struct headless_lms_server::prelude::web::Query

pub struct Query<T>(pub T);

Extract typed information from the request’s query.

To extract typed data from the URL query string, the inner type T must implement the DeserializeOwned trait.

Use QueryConfig to configure extraction process.

Panics

A query string consists of unordered key=value pairs, therefore it cannot be decoded into any type which depends upon data ordering (eg. tuples). Trying to do so will result in a panic.

Examples

use actix_web::{get, web};
use serde::Deserialize;

#[derive(Debug, Deserialize)]
pub enum ResponseType {
   Token,
   Code
}

#[derive(Debug, Deserialize)]
pub struct AuthRequest {
   id: u64,
   response_type: ResponseType,
}

// Deserialize `AuthRequest` struct from query string.
// This handler gets called only if the request's query parameters contain both fields.
// A valid request path for this handler would be `/?id=64&response_type=Code"`.
#[get("/")]
async fn index(info: web::Query<AuthRequest>) -> String {
    format!("Authorization request for id={} and type={:?}!", info.id, info.response_type)
}

// To access the entire underlying query struct, use `.into_inner()`.
#[get("/debug1")]
async fn debug1(info: web::Query<AuthRequest>) -> String {
    dbg!("Authorization object = {:?}", info.into_inner());
    "OK".to_string()
}

// Or use destructuring, which is equivalent to `.into_inner()`.
#[get("/debug2")]
async fn debug2(web::Query(info): web::Query<AuthRequest>) -> String {
    dbg!("Authorization object = {:?}", info);
    "OK".to_string()
}

Tuple Fields
0: T
Implementations

impl<T> Query<T>

pub fn into_inner(self) -> T

Unwrap into inner T value.

impl<T> Query<T>where T: DeserializeOwned,

pub fn from_query(query_str: &str) -> Result<Query<T>, QueryPayloadError>

Deserialize a T from the URL encoded query parameter string.

let numbers = Query::<HashMap<String, u32>>::from_query("one=1&two=2").unwrap();
assert_eq!(numbers.get("one"), Some(&1));
assert_eq!(numbers.get("two"), Some(&2));
assert!(numbers.get("three").is_none());