Struct headless_lms_server::prelude::web::ServiceConfig

pub struct ServiceConfig {
    pub(crate) services: Vec<Box<dyn AppServiceFactory>>,
    pub(crate) external: Vec<ResourceDef>,
    pub(crate) app_data: Extensions,
    pub(crate) default: Option<Rc<BoxServiceFactory<(), ServiceRequest, ServiceResponse, Error, ()>>>,
}

Enables parts of app configuration to be declared separately from the app itself. Helpful for modularizing large applications.

Merge a ServiceConfig into an app using App::configure. Scope and resources services have similar methods.

use actix_web::{web, App, HttpResponse};

// this function could be located in different module
fn config(cfg: &mut web::ServiceConfig) {
    cfg.service(web::resource("/test")
        .route(web::get().to(|| HttpResponse::Ok()))
        .route(web::head().to(|| HttpResponse::MethodNotAllowed()))
    );
}

// merge `/test` routes from config function to App
App::new().configure(config);

Fields
services: Vec<Box<dyn AppServiceFactory>>external: Vec<ResourceDef>app_data: Extensionsdefault: Option<Rc<BoxServiceFactory<(), ServiceRequest, ServiceResponse, Error, ()>>>
Implementations

impl ServiceConfig

pub fn data<U>(&mut self, data: U) -> &mut ServiceConfig
where U: 'static,

๐Deprecated since 4.0.0: Use .app_data(Data::new(val)) instead.

Add shared app data item.

Counterpart to App::data().

pub fn app_data<U>(&mut self, ext: U) -> &mut ServiceConfig
where U: 'static,

Add arbitrary app data item.

Counterpart to App::app_data().

pub fn default_service<F, U>(&mut self, f: F) -> &mut ServiceConfig
where F: IntoServiceFactory<U, ServiceRequest>, U: ServiceFactory<ServiceRequest, Config = (), Response = ServiceResponse, Error = Error> + 'static, <U as ServiceFactory<ServiceRequest>>::InitError: Debug,

Default service to be used if no matching resource could be found.

Counterpart to App::default_service().

pub fn configure<F>(&mut self, f: F) -> &mut ServiceConfig
where F: FnOnce(&mut ServiceConfig),

Run external configuration as part of the application building process

Counterpart to App::configure() that allows for easy nesting.

pub fn route(&mut self, path: &str, route: Route) -> &mut ServiceConfig

Configure route for a specific path.

Counterpart to App::route().

pub fn service<F>(&mut self, factory: F) -> &mut ServiceConfig
where F: HttpServiceFactory + 'static,

Register HTTP service factory.

Counterpart to App::service().

pub fn external_resource<N, U>(&mut self, name: N, url: U) -> &mut ServiceConfig
where N: AsRef<str>, U: AsRef<str>,

Register an external resource.

External resources are useful for URL generation purposes only and are never considered for matching at request time. Calls to HttpRequest::url_for() will work as expected.

Counterpart to App::external_resource().