Module headless_lms_server::programs::mailchimp_syncer

Structs

FieldSchema 🔒

MailchimpField 🔒

SyncerConfig 🔒

Structure to hold the configuration settings, such as the database URL.

Constants

BATCH_SIZE 🔒

FIELDS_EXCLUDED_FROM_REMOVING 🔒

These fields are excluded from removing all fields that are not in the schema

PRINT_STILL_RUNNING_MESSAGE_TICKS_THRESHOLD 🔒

PROCESS_UNSUBSCRIBES_INTERVAL_SECS 🔒

REMOVE_UNSUPPORTED_FIELDS 🔒

REQUIRED_FIELDS 🔒

SYNC_INTERVAL_SECS 🔒

Functions

add_field_to_mailchimp 🔒

Adds a new merge field to the Mailchimp list.

ensure_mailchimp_schema 🔒

Ensures the Mailchimp schema is up to date, adding required fields and removing any extra ones.

fetch_current_mailchimp_fields 🔒

Fetches the current merge fields from the Mailchimp list schema.

fetch_unsubscribed_users_from_mailchimp_in_chunks 🔒

Fetches data from Mailchimp in chunks.

initialize_configuration 🔒

Initializes and returns configuration settings (database URL).

initialize_database_pool 🔒

Initializes the PostgreSQL connection pool from the provided database URL.

initialize_environment 🔒

Initializes environment variables, logging, and tracing setup.

main

The main function that initializes environment variables, config, and sync process.

process_unsubscribed_users_from_mailchimp 🔒

remove_field_from_mailchimp 🔒

Removes a merge field from the Mailchimp list by with a field ID.

send_users_to_mailchimp

Sends a batch of users to Mailchimp for synchronization.

sync_contacts 🔒

Synchronizes the user contacts with Mailchimp. Added a boolean flag to determine whether to process unsubscribes.

sync_tags_from_mailchimp

Fetch tags from mailchimp and sync the changes to the database

update_emails_in_mailchimp 🔒

Updates the email addresses of multiple users in a Mailchimp mailing list.